An awful hack to create a target
If the target orocos-bfl does not exist, catkin_make may not succeed if the library
has never ben built before. It has no target to build to create the library, so
rather than relying on the package dependencies it just fails. This adds a dummy
target which is overwritten at the install stage by the correct thing.